+++ chef-repo/cookbooks/yum_repo/metadata.rb
@@ -1,7 +1,7 @@
-name             'yum_repository'
-maintainer       'YOUR_COMPANY_NAME'
-maintainer_email 'YOUR_EMAIL'
+name             'yum_repo'
+maintainer       'Cognizant Technology Solutions'
+maintainer_email 'suhail.ahmed@cognizant.com'
 license          'All rights reserved'
-description      'Installs/Configures yum_repository'
+description      'Installs/Configures yum_repo'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'