+++ chef-repo/cookbooks/yum_repo/recipes/default.rb
@@ -1,14 +1,13 @@
 #
-# Cookbook Name:: yum_repository
+# Cookbook Name:: yum_repo
 # Recipe:: default
 #
 # Copyright 2017, YOUR_COMPANY_NAME
 #
 # All rights reserved - Do Not Redistribute
 #
-
 template '/etc/yum.repos.d/microsoft.repo' do
-	source 'microsoft_repo.erb'
+	source 'redhat_microsoft_repo.erb'
 	owner 'root'
 	group 'root'
 	mode '0644'