+++ chef-repo/cookbooks/yum_repo/CHANGELOG.md
@@ -1,9 +1,9 @@
-# yum_repository CHANGELOG
+# yum_repo CHANGELOG
 
-This file is used to list changes made in each version of the yum_repository cookbook.
+This file is used to list changes made in each version of the yum_repo cookbook.
 
 ## 0.1.0
-- [your_name] - Initial release of yum_repository
+- [your_name] - Initial release of yum_repo
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.